Keep old review withdraw dialog for easier review
