+++ webapp/app/[locale]/tunnel/_components/claim.tsx
@@ -30,7 +30,7 @@ import {
 } from '../_hooks/useTunnelState'
 
 import { ReviewBtcDeposit } from './reviewOperation/reviewBtcDeposit'
-import { ReviewEvmWithdrawal } from './reviewOperation/reviewEvmWithdrawal'
+import { ReviewEvmWithdrawal } from './reviewOperation/reviewEvmWithdrawalOld'
 
 const EvmSubmitButton = function ({
   claimTxHash: inMemoryClaimTxHash,
